Extending from Controller class
We can extend our controllers from Controller class. This class provide us of several utilities in order to access to request and manipulate the response, among other things. The controllers that extends from Controller can have methods returning IActionResult objects. This is a important concept. When a method controller returns IActionResult, MVC don't send the response to the client, this responsability is done in other pipeline of the middleware which knows what have to do with that IActionResult.

There is several implmentations from IActionResult.