Add model to response
Here, instead of return a string, I return an object, the model. I do that returning a ObjectResult object, but, here, I don't say what format I am going to return that object: JSON, XML, text plain, etc... it will be the responsability in other place in pipeline. But, by default, the response is done in JSON.