View storeroom details
In this commit I have implemented a controller that handles a parameter, in this case, the parameter goes in the request's path. I have configured the routing adding /{id?}. With ? character, I am telling to MVC that is optional.

I can handle requests as /home/details/2 or /home/details?id=2